Extract REST related stuff from main provider impl module into a new REST module and related SPI module
