+++ modules/apps/oauth2-provider/oauth2-provider-rest-spi/bnd.bnd
@@ -2,4 +2,5 @@ Bundle-Name: Liferay OAuth2 Provider REST SPI
 Bundle-SymbolicName: com.liferay.oauth2.provider.rest.spi
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.oauth2.provider.rest.spi
+	com.liferay.oauth2.provider.rest.scopechecker.spi,\
+	com.liferay.oauth2.provider.rest.tokenprovider.spi
