+++ modules/apps/oauth2-provider/oauth2-provider-rest-spi/build.gradle
@@ -2,5 +2,7 @@ sourceCompatibility = "1.8"
 targetCompatibility = "1.8"
 
 dependencies {
+	compileOnly group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.0.1"
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-api")
+	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-api")
 }
