+++ modules/apps/oauth2-provider/oauth2-provider-rest/src/main/java/com/liferay/oauth2/provider/rest/requestscopechecker/AnnotationRequestScopeChecker.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.oauth2.provider.scopes.impl.requestscopechecker;
+package com.liferay.oauth2.provider.rest.requestscopechecker;
 
+import com.liferay.oauth2.provider.rest.scopechecker.spi.RequestScopeChecker;
 import com.liferay.oauth2.provider.scopes.api.RequiresScope;
 import com.liferay.oauth2.provider.scopes.api.ScopeChecker;
-import com.liferay.oauth2.provider.scopes.spi.RequestScopeChecker;
 import com.liferay.portal.kernel.util.MapUtil;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
