+++ modules/apps/oauth2-provider/oauth2-provider-scopes-impl/build.gradle
@@ -20,6 +20,7 @@ dependencies {
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-api")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-liferay-api")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-spi")
+	compileOnly project(":apps:oauth2-provider:oauth2-provider-rest-spi")
 
 	testCompile group: "junit", name: "junit", version: "4.12"
 }
