+++ modules/apps/oauth2-provider/oauth2-provider-scopes-impl/src/main/java/com/liferay/oauth2/provider/scopes/impl/feature/LiferayOAuth2OSGiFeature.java
@@ -16,10 +16,10 @@ package com.liferay.oauth2.provider.scopes.impl.feature;
 
 import com.liferay.oauth2.provider.scopes.impl.jaxrs.CompanyRetrieverContainerRequestFilter;
 import com.liferay.oauth2.provider.scopes.impl.jaxrs.RunnableExecutorContainerResponseFilter;
+import com.liferay.oauth2.provider.rest.scopechecker.spi.RequestScopeChecker;
 import com.liferay.oauth2.provider.scopes.api.ScopeChecker;
 import com.liferay.oauth2.provider.scopes.impl.jaxrs.ScopedRequestScopeChecker;
 import com.liferay.oauth2.provider.scopes.spi.ApplicationDescriptor;
-import com.liferay.oauth2.provider.scopes.spi.RequestScopeChecker;
 import com.liferay.oauth2.provider.scopes.spi.ScopeDescriptor;
 import com.liferay.oauth2.provider.scopes.spi.ScopeFinder;
 import com.liferay.oauth2.provider.scopes.liferay.api.ScopeContext;
