+++ modules/apps/oauth2-provider/oauth2-provider-scopes-impl/src/main/java/com/liferay/oauth2/provider/scopes/impl/jaxrs/ScopedRequestScopeChecker.java
@@ -14,9 +14,9 @@
 
 package com.liferay.oauth2.provider.scopes.impl.jaxrs;
 
+import com.liferay.oauth2.provider.rest.scopechecker.spi.RequestScopeChecker;
 import com.liferay.oauth2.provider.scopes.api.ScopeChecker;
 import com.liferay.oauth2.provider.scopes.liferay.api.ScopedServiceTrackerMap;
-import com.liferay.oauth2.provider.scopes.spi.RequestScopeChecker;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.util.PortalUtil;
