+++ modules/apps/oauth2-provider/oauth2-provider-scopes-impl/src/test/java/com/liferay/oauth2/provider/scopes/impl/HttpMethodRequestScopeCheckerTest.java
@@ -19,8 +19,8 @@ import static junit.framework.TestCase.assertTrue;
 import javax.ws.rs.container.ResourceInfo;
 import javax.ws.rs.core.Request;
 
+import com.liferay.oauth2.provider.rest.scopechecker.spi.RequestScopeChecker;
 import com.liferay.oauth2.provider.scopes.impl.requestscopechecker.HttpMethodRequestScopeChecker;
-import com.liferay.oauth2.provider.scopes.spi.RequestScopeChecker;
 
 import org.junit.Before;
 import org.junit.Test;
